requirements.txt
astral
gpiozero
RPi.GPIO
pyserial
pytz
supervisor
tweepy
wiringpi2
